Backend/requirements.txt
# requirements.txt

# Web framework
Flask==2.3.2
flask-cors==3.0.10

# For handling HTTP requests in the frontend
requests==2.31.0

# Hugging Face Transformers library for working with StarCoder
transformers==4.30.2

# PyTorch for the underlying machine learning operations
torch==2.0.1

# For efficient tensor operations on CUDA-enabled GPUs
accelerate==0.20.3

# For environment variable management
python-dotenv==1.0.0

# For production WSGI server
gunicorn==20.1.0

# Optional: for better performance with PyTorch
ninja==1.11.1

# trying to offload weights to disk
safetensors==0.3.1

#token=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